Fixed an issue with localization of class names (class filters)
Issue:
The base classes were hardcoded into Item Assistant to help remedy a scaling issue with 4K monitors.
When these classes were added as hardcoded, localization support was incorrectly not added for these checkboxes.

Furthermore, there was a bug causing the existing classes to be added again (uniqueness was based on name, not the class tag itself)

Fix:
Removed the IA specific class tags (iatag_ui_CLASSNAME) as they are redundant and missing in some translations, and ensured that the translated text tag is used for hardcoded checkboxes.