Should now wait 6 seconds before starting the item listener thread upon injection. Will also no longer inject into the game until the game is actually running.
This should hopefully prevent some of the most prevalent menu crashes.
It does not however address the issue of going back to the menu later on, and loading a new characters.

I suspect the main issue here is that the game goes into a "ready" state while still loading, and crashes when attempting to parse a bunch of item tooltips out of the blue.